Introduce EnvironmentAdapter in place of action and state mapper functions (#35)
* Introduce EnvironmentAdapter in place of action and state mapper functions

We already had action and state conversion functions for conversions
between the environment representation and the ACS2 representation.
Whenever the environment representation of actions or states was
incompatible with what ACS2 received and produced, adapter functions
needed to be used.

However, we only had uni-directional converters:
  * env state -> ACS2 state
  * ACS2 action -> env action

The goal of this commit is to add the reverse converters, i.e.:
  * env action -> ACS2 action
  * ACS2 state -> env state
so that, e.g., Classifier.__str__() can be implemented properly.

For instance, this allows Classifier.__str__() to print 'N', 'W', 'SE',
etc. in the action part instead of 0, 6, 3, etc. if the proper
environment adapter is provided.

Notebooks not updated (yet).

RACS not touched.

* fixup! Introduce EnvironmentAdapter in place of action and state mapper functions

naming